+++ src/app/app.component.html
@@ -1,1 +1,1 @@
-<router-outlet></router-outlet>
+<app-interactive></app-interactive>
