--- src/app/interactive-card/back-card/back-card.compoent.css
@@ -1,27 +1,0 @@
-.name-card-back {
-  padding-left: 90%;
-}
-
-.card-back {
-  position: absolute;
-  top: 65%;
-  left: 25%;
-  background-image: url("../../../assets/images/bg-card-back.png");
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  height: 264px;
-  width: 430px;
-  border-radius: 16px;
-  transform: translate(-50%, -50%);
-  background-color: #ffffff;
-  border: 1px solid #dcdcdc;
-  border-radius: 8px;
-  padding: 16px;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.card-redit-back {
-  padding-top: 22%;
-  color: aliceblue;
-}
