--- src/app/interactive-card/front-card/front-card.component.css
@@ -1,35 +1,0 @@
-.card-front {
-  position: absolute;
-  top: 26%;
-  left: 25%;
-  background-image: url("../../../assets/images/bg-card-front.png");
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  height: 264px;
-  width: 430px;
-  border-radius: 16px;
-  transform: translate(-50%, -50%);
-  background-color: #ffffff;
-  border: 1px solid #dcdcdc;
-  border-radius: 8px;
-  padding: 16px;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.name-data-card-front {
-  font-size: 18px;
-}
-
-.name-card-front {
-  padding-right: 52%;
-}
-
-.number-front-card {
-  font-size: 48px;
-}
-
-.card-redit-front {
-  padding-top: 30%;
-  color: aliceblue;
-}
