+++ src/app/interactive-card/interactive-card.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule, NgIf } from "@angular/common";
 import { NgModule } from "@angular/core";
-import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { FormsModule } from "@angular/forms";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatInputModule } from '@angular/material/input';
 import { BrowserModule } from "@angular/platform-browser";
@@ -15,18 +15,9 @@ import { NzMenuModule } from "ng-zorro-antd/menu";
 import { NzInputModule } from "ng-zorro-antd/input";
 import { NzInputNumberModule } from "ng-zorro-antd/input-number";
 import { NzMessageService } from 'ng-zorro-antd/message';
-import { FrontCardComponent } from "./front-card/front-card.component";
-import { InteractiveCardComponent } from "./interactive-card.component";
-import { NzButtonModule } from "ng-zorro-antd/button";
-import { BackCardCompoent } from "./back-card/back-card.component";
-import { FormCardComponent } from "./form-card/form-card.component";
 
 @NgModule({
     declarations: [
-        InteractiveCardComponent,
-        FrontCardComponent,
-        BackCardCompoent,
-        FormCardComponent
     ],
     imports: [
         CommonModule,
@@ -47,14 +38,11 @@ import { FormCardComponent } from "./form-card/form-card.component";
         NzCardModule,
         NzFormModule,
         NzInputModule,
-        NzInputNumberModule,
-        ReactiveFormsModule,
-        NzButtonModule
+        NzInputNumberModule
     ],
     providers: [
         NzMessageService
     ],
-
 })
 
 export class InteractiveModule { }
